minor bug fixes and improvements
Added a cache breaker so that different words that using the same image
load correctly.

Fix null reference error in SearchContainer.js

Auto focused the search input on page load.

Aligned the spinner to the center of the box.

Disable HTTPS redirect when on localhost.

Improved description for home page input box